Good advise and way forward from Paul:
As a suggestion would it not be better to keep the `edrawings-viewer-2019`
pointing to the 2019 version, the `edrawings-viewer-2018` package pointing to
the 2018 version and update the `edrawings-viewer` package to point to the
latest (`edrawings-viewer-2019` at the moment)? That way those using the
`edrawings-viewer` package will always have the most up to date version and
those wanting to stick with the 2018 version can just use the
`edrawings-viewer-2018` package?

The risk of just having the one `edrawings-viewer` package pointing to the
latest and the others taking that as a dependency means those on the 2018
version suddenly get upgraded to the latest with no say so.